requirements.txt
streamlit==1.51.0
pyalex==0.19
psycopg2-binary==2.9.9
python-dotenv==1.0.0

